# docs/swagger.yaml
openapi: 3.0.0
paths:
  /api/metrics:
    get:
      operationId: MetricsController_scrape
      summary: Prometheus metrics exposition
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - metrics
  /api/auth/register:
    post:
      operationId: AuthController_register
      summary: Register a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User created; returns token + user.
        '409':
          description: Email or username already in use.
      tags: &ref_0
        - auth
  /api/auth/login:
    post:
      operationId: AuthController_login
      summary: Log in with email and password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Returns token + user.
        '401':
          description: Invalid credentials or unverified email.
        '403':
          description: Account suspended.
      tags: *ref_0
  /api/auth/send-code:
    post:
      operationId: AuthController_sendVerificationCode
      summary: Send a UCF email verification code
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeDto'
      responses:
        '200':
          description: Verification code sent.
        '400':
          description: Invalid UCF email.
        '429':
          description: Requested again within the cooldown.
      tags: *ref_0
  /api/auth/verify-code:
    post:
      operationId: AuthController_verifyCode
      summary: Verify the UCF email code
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeDto'
      responses:
        '200':
          description: Email verified; account activated.
        '400':
          description: Invalid or expired code.
      tags: *ref_0
  /api/auth/forgot-password:
    post:
      operationId: AuthController_forgotPassword
      summary: Email a single-use password-reset link
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: Reset link sent if the email exists.
        '502':
          description: The reset email could not be sent.
      tags: *ref_0
  /api/auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Redeem a reset link and set a new password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset successfully.
        '400':
          description: Invalid or expired token, or a rejected password.
      tags: *ref_0
  /api/auth/change-password:
    patch:
      operationId: AuthController_changePassword
      summary: Change the authenticated user's password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Password changed successfully.
        '400':
          description: Current password incorrect, or new password rejected.
      tags: *ref_0
      security:
        - bearer: []
  /api/users/me:
    get:
      operationId: UsersController_getMe
      summary: Get the authenticated user's profile
      parameters: []
      responses:
        '200':
          description: The current user profile.
        '401':
          description: Missing or invalid token.
      tags: &ref_1
        - users
      security: &ref_2
        - bearer: []
    patch:
      operationId: UsersController_updateMe
      summary: Update the authenticated user's profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Updated profile.
        '409':
          description: Username already in use.
      tags: *ref_1
      security: *ref_2
    delete:
      operationId: UsersController_deleteMe
      summary: Soft-delete the authenticated account
      parameters: []
      responses:
        '204':
          description: Account soft-deleted.
      tags: *ref_1
      security: *ref_2
  /api/users/me/avatar:
    put:
      operationId: UsersController_uploadAvatar
      summary: Upload or replace the authenticated user's picture
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WebP image, 5 MB max.
      responses:
        '200':
          description: Updated profile with picture path.
        '400':
          description: Missing file or unsupported type.
        '413':
          description: Image exceeds the 5 MB limit.
      tags: *ref_1
      security: *ref_2
    delete:
      operationId: UsersController_removeAvatar
      summary: Remove the authenticated user's picture
      parameters: []
      responses:
        '200':
          description: Picture removed.
      tags: *ref_1
      security: *ref_2
  /api/users/username-available:
    get:
      operationId: UsersController_usernameAvailable
      parameters:
        - name: username
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
      security: *ref_2
  /api/users/{id}:
    get:
      operationId: UsersController_getPublicProfile
      summary: View another player's public profile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Public profile fields only.
        '404':
          description: User not found.
      tags: *ref_1
      security: *ref_2
  /api/games:
    post:
      operationId: GamesController_create
      summary: Host a new game
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameDto'
      responses:
        '201':
          description: The created game.
        '400':
          description: Invalid payload.
      tags: &ref_3
        - games
      security: &ref_4
        - bearer: []
    get:
      operationId: GamesController_findMany
      summary: Discover games
      parameters:
        - name: sport
          required: false
          in: query
          schema:
            example: soccer
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - open
              - confirmed
              - locked
              - completed
              - cancelled
            type: string
        - name: upcoming
          required: false
          in: query
          description: Only games whose start_time is in the future.
          schema:
            default: true
            type: boolean
      responses:
        '200':
          description: Matching games.
      tags: *ref_3
      security: *ref_4
  /api/games/mine:
    get:
      operationId: GamesController_findMine
      summary: Games the caller hosts or plays in
      parameters:
        - name: role
          required: false
          in: query
          description: 'Only games you host or only games you play in. Default: both.'
          schema:
            enum:
              - hosting
              - playing
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - open
              - confirmed
              - locked
              - completed
              - cancelled
            type: string
      responses:
        '200':
          description: The caller's games.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}:
    get:
      operationId: GamesController_findOne
      summary: Get a single game
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The game.
        '404':
          description: Game not found.
      tags: *ref_3
      security: *ref_4
    patch:
      operationId: GamesController_update
      summary: Edit a game (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameDto'
      responses:
        '200':
          description: The updated game.
        '403':
          description: Only the host can edit.
      tags: *ref_3
      security: *ref_4
    delete:
      operationId: GamesController_remove
      summary: Delete a game (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The game was deleted.
        '403':
          description: Only the host can delete.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/photo:
    put:
      operationId: GamesController_uploadPhoto
      summary: Upload or replace a game's banner (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WebP image, 5 MB max.
      responses:
        '200':
          description: Updated game with banner path.
        '400':
          description: Missing file or unsupported type.
        '403':
          description: Only the host can set the banner.
        '413':
          description: Image exceeds the 5 MB limit.
      tags: *ref_3
      security: *ref_4
    delete:
      operationId: GamesController_removePhoto
      summary: >-
        Remove a game's banner, reverting to the sport's stock default (host
        only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Updated game with the stock banner restored.
        '403':
          description: Only the host can remove the banner.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/join:
    post:
      operationId: GamesController_join
      summary: Join a game, optionally for a party of more than one
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGameDto'
      responses:
        '200':
          description: The joined game.
        '400':
          description: >-
            Game full/started, already joined, or the party doesn't fit the
            remaining spots.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/leave:
    post:
      operationId: GamesController_leave
      summary: Leave a game
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The updated game.
        '400':
          description: Host cannot leave / not on roster.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/guests:
    post:
      operationId: GamesController_addGuest
      summary: Add a guest player to the roster (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitialPlayerDto'
      responses:
        '201':
          description: The updated game.
        '400':
          description: Game full/started/terminal, or invalid guest.
        '403':
          description: Only the host can add guests.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/guests/{index}:
    delete:
      operationId: GamesController_removeGuest
      summary: Remove a guest from the roster by index (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: index
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: The updated game.
        '400':
          description: No guest at that index.
        '403':
          description: Only the host can remove guests.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/position:
    patch:
      operationId: GamesController_setMyPosition
      summary: Set your own position on a game you're on
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPositionDto'
      responses:
        '200':
          description: The updated game.
        '400':
          description: Not on the roster, or the game is terminal.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/cancel:
    post:
      operationId: GamesController_cancel
      summary: Cancel a game (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The cancelled game.
        '403':
          description: Only the host can cancel.
      tags: *ref_3
      security: *ref_4
  /api/games/{id}/complete:
    post:
      operationId: GamesController_complete
      summary: Mark a game completed (host only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The completed game.
        '403':
          description: Only the host can complete.
      tags: *ref_3
      security: *ref_4
  /api/notifications/device-token:
    post:
      operationId: NotificationsController_registerDevice
      summary: Register a device token for push notifications
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceDto'
      responses:
        '201':
          description: Device token registered.
      tags: &ref_5
        - notifications
  /api/notifications:
    get:
      operationId: NotificationsController_getMyNotifications
      summary: Get all notifications for the current user
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_5
  /api/notifications/{id}/read:
    patch:
      operationId: NotificationsController_markRead
      summary: Mark a notification as read
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_5
  /api/notifications/read-all:
    patch:
      operationId: NotificationsController_markAllRead
      summary: Mark all notifications as read
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_5
info:
  title: SquadUp API
  description: Authentication, user, and games endpoints for the SquadUp sports social app.
  version: 0.1.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterDto:
      type: object
      properties:
        first_name:
          type: string
          example: Alex
        last_name:
          type: string
          example: Rivera
        username:
          type: string
          example: alex_r
          description: Unique public handle
        email:
          type: string
          example: alex@ucf.edu
          description: Must be a UCF email address (@ucf.edu).
        password:
          type: string
          example: Passw0rd!
          minLength: 8
          maxLength: 20
          description: >-
            Password: 8–20 characters, with at least one uppercase letter, one
            lowercase letter, one number, and one symbol.
      required:
        - first_name
        - last_name
        - username
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: alex@ucf.edu
        password:
          type: string
          example: password123
      required:
        - email
        - password
    SendCodeDto:
      type: object
      properties:
        email:
          type: string
          example: ta326121@ucf.edu
      required:
        - email
    VerifyCodeDto:
      type: object
      properties:
        email:
          type: string
          example: ta326121@ucf.edu
        code:
          type: string
          example: '123456'
      required:
        - email
        - code
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: alex@ucf.edu
          description: Must be a UCF email address (@ucf.edu).
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          example: 3f1a…
          description: The token from the reset link. Single-use; expires in 1 hour.
        new_password:
          type: string
          example: Passw0rd!
          minLength: 8
          maxLength: 20
          description: >-
            Password: 8–20 characters, with at least one uppercase letter, one
            lowercase letter, one number, and one symbol.
      required:
        - token
        - new_password
    ChangePasswordDto:
      type: object
      properties:
        current_password:
          type: string
          example: CurrentPassw0rd!
        new_password:
          type: string
          example: N3wPassw0rd!
          minLength: 8
          maxLength: 20
          description: >-
            Password: 8–20 characters, with at least one uppercase letter, one
            lowercase letter, one number, and one symbol.
      required:
        - current_password
        - new_password
    UpdateProfileDto:
      type: object
      properties:
        first_name:
          type: string
          example: Alex
        last_name:
          type: string
          example: Rivera
        username:
          type: string
          example: alex_r
          description: Unique public handle
        preferred_positions:
          type: object
          description: Preferred position per sport (one each), keyed by sport.
          example:
            soccer: GK
            basketball: PG
          additionalProperties:
            type: string
    InitialPlayerDto:
      type: object
      properties:
        name:
          type: string
          example: Sam Lee
        position:
          type: string
          example: Goalkeeper
      required:
        - name
    CreateGameDto:
      type: object
      properties:
        sport:
          type: string
          example: soccer
        description:
          type: string
          example: Casual 5-a-side, all levels welcome.
        location:
          type: string
          example: North Campus Field 2
        start_time:
          type: string
          example: '2026-07-15T18:00:00.000Z'
          description: Kickoff time (ISO 8601). Must be in the future.
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.006
        min_players:
          type: number
          example: 6
          description: Roster size that confirms the game.
        max_players:
          type: number
          example: 10
          description: Roster size that locks the game.
        photo_url:
          type: string
          example: https://cdn.squadup.app/games/abc.jpg
          description: >-
            Banner image URL. When omitted, defaults to the sport's stock
            banner.
        players:
          example:
            - name: Alex Rivera
              position: Forward
            - name: Sam Lee
          description: >-
            Initial roster of players to pre-add. These are guests who may not
            have an account; each counts toward min/max like any player and may
            carry an optional position. The host is added automatically and
            should not be listed here.
          type: array
          items:
            $ref: '#/components/schemas/InitialPlayerDto'
        host_position:
          type: string
          example: Midfielder
          description: >-
            The host's own (optional, sport-specific) position on the roster.
            The host is auto-added as a player, so this records where they play.
      required:
        - sport
        - location
        - start_time
        - latitude
        - longitude
        - min_players
        - max_players
    UpdateGameDto:
      type: object
      properties:
        sport:
          type: string
          example: soccer
        description:
          type: string
          example: Casual 5-a-side, all levels welcome.
        location:
          type: string
          example: North Campus Field 2
        start_time:
          type: string
          example: '2026-07-15T18:00:00.000Z'
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.006
        min_players:
          type: number
          example: 6
        max_players:
          type: number
          example: 10
        photo_url:
          type: string
          example: https://cdn.squadup.app/games/abc.jpg
          description: >-
            Banner image URL. Changing the sport re-points a stock banner at the
            new sport; a custom image is kept.
    JoinGameDto:
      type: object
      properties:
        party_size:
          type: number
          example: 3
          minimum: 1
          maximum: 20
          description: >-
            Total headcount this join represents, including the caller. Rejected
            if the group would exceed the remaining spots.
    SetPositionDto:
      type: object
      properties:
        position:
          type: string
          example: Midfielder
          description: Your (sport-specific, free-text) position. Empty/omitted clears it.
    RegisterDeviceDto:
      type: object
      properties:
        token:
          type: string
          example: fcm-device-token-here
        platform:
          type: string
          example: ios
      required:
        - token